carbon_spider/requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
pandas>=2.0
python-dateutil>=2.8
newspaper3k>=0.2.8
